requirements.txt
Django==2.2.7
django-webpack-loader==0.6.0
djangorestframework==3.10.3
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
